# provider-data-api/bundled-openapi-spec.yaml
openapi: 3.0.3
info:
  title: Provider API
  version: 1.0.0
  license:
    name: MIT
    url: https://github.com/ministryofjustice/laa-provider-data-platform/blob/main/LICENSE
  description: |
    API for managing Legal Service Providers, including Chambers, solicitors, and advocates.

    Unless explicitly stated otherwise, `officeGUID` values identify the provider-office link
    record, not the underlying office record. The underlying office GUID is mainly an internal
    identifier used to associate an office record with a provider-office link.
servers:
  - url: https://laa-data-provider-api-dev.apps.live.cloud-platform.service.justice.gov.uk
    description: Development
  - url: https://laa-data-provider-api-uat.apps.live.cloud-platform.service.justice.gov.uk
    description: UAT
  - url: https://laa-data-provider-api-staging.apps.live.cloud-platform.service.justice.gov.uk
    description: Staging
security:
  - ApiKeyAuth: []
  - AzureAD: []
paths:
...
  /novations:
    post:
      summary: Create a Novation
      security:
        - ApiKeyAuth: []
        - AzureAD: []
      operationId: createNovation
      tags:
        - Novations
      description: |
        Create a Novation record together with its initial predecessor/successor provider
        relationships. At least one relationship is required; each relationship requires a
        predecessor and successor provider firm and may optionally reference a Child Office.

        A Novation cannot be created directly with a status of Rescinded. Decision Date and
        Decision Reason requirements depend on the selected Novation Status (BR-39).
      parameters:
        - $ref: '#/components/parameters/TraceParentHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovationCreateV2'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/NovationCreateResponseV2'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: A referenced provider or office was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    get:
      summary: Retrieve Novations
      security:
        - ApiKeyAuth: []
        - AzureAD: []
      operationId: getNovations
      tags:
        - Novations
      description: |
        Retrieve Novations with optional filters.

        Not yet implemented; published as a contract-only stub.
      parameters:
        - $ref: '#/components/parameters/TraceParentHeader'
        - name: novationGUID
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
          description: Filter by Novation GUID (multi)
        - name: novationStatus
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/NovationStatusV2'
          description: Filter by Novation status (multi)
        - name: providerFirmGUID
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
          description: Filter by predecessor or successor provider firm GUID (multi)
        - $ref: '#/components/parameters/QPageNumber'
        - $ref: '#/components/parameters/QPageSize'
      responses:
        '200':
          description: List of Novations matching the filters
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    required:
                      - content
                      - metadata
                      - links
                    properties:
                      content:
                        type: array
                        items:
                          $ref: '#/components/schemas/NovationV2'
                      metadata:
                        $ref: '#/components/schemas/PaginatedSearchV2'
                      links:
                        $ref: '#/components/schemas/LinksV2'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /novations/{novationGUID}:
    get:
      summary: Retrieve a Novation
      security:
        - ApiKeyAuth: []
        - AzureAD: []
      operationId: getNovation
      tags:
        - Novations
      description: |
        Retrieve a Novation record and its supported predecessor/successor provider relationships.
        Supported relationships include Legal Organisation relationships and Child Office
        relationships where those Office links are available.

        This is a read-only operation and does not create, amend or delete Novation data or
        relationship data.
      parameters:
        - $ref: '#/components/parameters/TraceParentHeader'
        - $ref: '#/components/parameters/NovationGUID'
      responses:
        '200':
          description: Novation details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/NovationV2'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    patch:
      summary: Amend a Novation
      security:
        - ApiKeyAuth: []
        - AzureAD: []
      operationId: updateNovation
      tags:
        - Novations
      description: |
        Amend permitted Novation details or perform a permitted status transition.

        Not yet implemented; published as a contract-only stub for DSTEW-1977.
      parameters:
        - $ref: '#/components/parameters/TraceParentHeader'
        - $ref: '#/components/parameters/NovationGUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovationPatchV2'
      responses:
        '200':
          description: Novation updated successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/NovationV2'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
components:
  schemas:
    BaseEntityGUIDV2:
      type: string
      format: uuid
      description: GUID is the primary key for entity
      example: 123e4567-e89b-12d3-a456-426614174000
    BaseEntityV2:
      type: object
      required:
        - guid
        - version
        - createdBy
        - createdTimestamp
        - lastUpdatedBy
        - lastUpdatedTimestamp
      properties:
        guid:
          $ref: '#/components/schemas/BaseEntityGUIDV2'
        version:
          type: integer
          format: int64
          description: Record version (for audit and event concurrency)
          example: 12
        createdBy:
          type: string
          description: User ID from the auth token which resulted in creating this record
          example: john.smith.42
        createdTimestamp:
          type: string
          format: date-time
          example: '2025-09-25T09:30:00Z'
        lastUpdatedBy:
          type: string
          description: User ID from the auth token which resulted in updating this record last
          example: john.doe.24
        lastUpdatedTimestamp:
          type: string
          format: date-time
          example: '2025-09-26T12:30:00Z'
...
    BadRequestError:
      type: object
      properties:
        error:
          type: object
          properties:
            errorCode:
              type: string
              example: P00BR
    ForbiddenError:
      type: object
      properties:
        error:
          type: object
          properties:
            errorCode:
              type: string
              example: P00AD
    ConflictError:
      type: object
      properties:
        error:
          type: object
          properties:
            errorCode:
              type: string
              example: P00CV
    NotFoundError:
      type: object
      properties:
        error:
          type: object
          properties:
            errorCode:
              type: string
              example: P00NF
    InternalServerError:
      type: object
      properties:
        error:
          type: object
          properties:
            errorCode:
              type: string
              example: P00SE
    NovationTypeV2:
      type: string
      description: |
        Business classification of the Novation. The final permitted value set is still to be
        confirmed by the business, so this is intentionally not constrained by an enum yet.
      example: Merger
    NovationStatusV2:
      type: string
      enum:
        - Proposed
        - Approved
        - Approved with Conditions
        - Rejected
        - Withdrawn
        - Rescinded
    NovationCreateStatusV2:
      type: string
      description: Initial Novation status. A Novation cannot be created directly as Rescinded.
      enum:
        - Proposed
        - Approved
        - Approved with Conditions
        - Rejected
        - Withdrawn
    NovationCoreV2:
      type: object
      required:
        - novationType
        - novationEffectiveDate
      properties:
        novationType:
          $ref: '#/components/schemas/NovationTypeV2'
        novationEffectiveDate:
          type: string
          format: date
        novationStatus:
          $ref: '#/components/schemas/NovationStatusV2'
        decisionDate:
          type: string
          format: date
          nullable: true
          description: Required when the Novation has moved out of Proposed.
        decisionReason:
          type: string
          nullable: true
          description: Required for decision statuses as defined by BR-39.
        decisionBy:
          type: string
          nullable: true
          readOnly: true
          description: User ID from the authenticated request which moved the Novation out of Proposed.
        rescindedDate:
          type: string
          format: date
          nullable: true
          description: Required when the status is Rescinded.
        rescindedReason:
          type: string
          nullable: true
          description: Required when the status is Rescinded.
        rescindedBy:
          type: string
          nullable: true
          readOnly: true
          description: User ID from the authenticated request which rescinded the Novation.
        driverForNovation:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    NovationCreateV2:
      type: object
      required:
        - novationType
        - novationEffectiveDate
        - relationships
      properties:
        novationType:
          $ref: '#/components/schemas/NovationTypeV2'
        novationEffectiveDate:
          type: string
          format: date
        novationStatus:
          $ref: '#/components/schemas/NovationCreateStatusV2'
        decisionDate:
          type: string
          format: date
          nullable: true
        decisionReason:
          type: string
          nullable: true
          description: Decision reason required according to the initial non-Proposed status.
        driverForNovation:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        relationships:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/NovationRelationshipCreateV2'
          description: |
            Predecessor/successor relationships for this Novation. Multiple entries support
            many-to-one and one-to-many provider changes.
    NovationV2:
      allOf:
        - $ref: '#/components/schemas/BaseEntityV2'
        - $ref: '#/components/schemas/NovationCoreV2'
        - type: object
          required:
            - relationships
          properties:
            relationships:
              type: array
              items:
                $ref: '#/components/schemas/NovationRelationshipV2'
    NovationCreateResponseV2:
      type: object
      required:
        - novationGUID
        - novationRelationshipGUIDs
      properties:
        novationGUID:
          $ref: '#/components/schemas/BaseEntityGUIDV2'
        novationRelationshipGUIDs:
          type: array
          items:
            $ref: '#/components/schemas/BaseEntityGUIDV2'
    NovationRelationshipCreateV2:
      type: object
      required:
        - previousProviderFirmGUID
        - newProviderFirmGUID
      properties:
        previousProviderFirmGUID:
          $ref: '#/components/schemas/BaseEntityGUIDV2'
        newProviderFirmGUID:
          $ref: '#/components/schemas/BaseEntityGUIDV2'
        previousOfficeGUID:
          type: string
          format: uuid
          nullable: true
          description: GUID of the provider-office link, not the underlying physical office row.
        newOfficeGUID:
          type: string
          format: uuid
          nullable: true
          description: GUID of the provider-office link, not the underlying physical office row.
        previousContractGUID:
          type: string
          format: uuid
          nullable: true
        newContractGUID:
          type: string
          format: uuid
          nullable: true
        previousScheduleGUID:
          type: string
          format: uuid
          nullable: true
        newScheduleGUID:
          type: string
          format: uuid
          nullable: true
        notes:
          type: string
          nullable: true
      description: |
        One predecessor/successor mapping belonging to a Novation. Previous/new Contract GUIDs
        must be supplied together when a Contract relationship is recorded. Previous/new Schedule
        GUIDs must be supplied together when a Schedule relationship is recorded. Referenced
        Offices, Contracts and Schedules must be consistent with the referenced provider firms;
        those cross-entity rules are enforced by the service.
    NovationRelationshipV2:
      allOf:
        - $ref: '#/components/schemas/BaseEntityV2'
        - $ref: '#/components/schemas/NovationRelationshipCreateV2'
        - type: object
          required:
            - novationGUID
          properties:
            novationGUID:
              $ref: '#/components/schemas/BaseEntityGUIDV2'
    NovationPatchV2:
      type: object
      x-class-extra-annotation: '@uk.gov.justice.laa.providerdata.annotation.RejectProperties( {"guid", "relationships", "decisionBy", "rescindedBy", "createdBy", "createdTimestamp", "lastUpdatedBy", "lastUpdatedTimestamp"})'
      properties:
        novationType:
          $ref: '#/components/schemas/NovationTypeV2'
        novationEffectiveDate:
          type: string
          format: date
        novationStatus:
          $ref: '#/components/schemas/NovationStatusV2'
        decisionDate:
          type: string
          format: date
          nullable: true
        decisionReason:
          type: string
          nullable: true
          description: Original decision reason; must not be overwritten by a later rescission.
        rescindedDate:
          type: string
          format: date
          nullable: true
          description: Required when transitioning an Approved or Approved with Conditions Novation to Rescinded.
        rescindedReason:
          type: string
          nullable: true
          description: Required when transitioning an Approved or Approved with Conditions Novation to Rescinded.
        driverForNovation:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    PaginatedSearchV2:
      type: object
      required:
        - searchCriteria
        - pagination
      properties:
        searchCriteria:
          $ref: '#/components/schemas/SearchCriteriaV2'
        pagination:
          $ref: '#/components/schemas/PaginationV2'
        sort:
          $ref: '#/components/schemas/SortV2'
    SearchCriteriaV2:
      type: object
      properties:
        criteria:
          type: array
          items:
            $ref: '#/components/schemas/SearchCriterionV2'
    SearchCriterionV2:
      type: object
      properties:
        filter:
          type: string
          example: guid
          description: Filter key
        values:
          type: array
          items:
            type: string
            example: 123e4567-e89b-12d3-a456-426614174000
            description: Filter value(s) used for this search
    PaginationV2:
      type: object
      properties:
        currentPage:
          type: integer
          example: 2
          description: Page index denoting the current page index for response, starting from 0
        pageSize:
          type: integer
          example: 10
          description: Page size used for this search
        totalPages:
          type: integer
          example: 5
          description: Total number of pages
        totalItems:
          type: integer
          format: int64
          example: 123
          description: Total number of records for given search request
    SortV2:
      type: object
      properties:
        field:
          type: string
          example: sortByOption
        direction:
          type: string
          enum:
            - asc
            - desc
          example: asc
    LinksV2:
      type: object
      properties:
        self:
          type: string
          example: https://example.com/example?page=2&pageSize=10
          format: uri
        next:
          type: string
          example: https://example.com/example?page=3&pageSize=10
          format: uri
        prev:
          type: string
          example: https://example.com/example?page=1&pageSize=10
          format: uri
        first:
          type: string
          example: https://example.com/example?page=1&pageSize=10
          format: uri
        last:
          type: string
          example: https://example.com/example?page=5&pageSize=10
          format: uri
...
  parameters:
...
    TraceParentHeader:
      name: traceparent
      in: header
      required: false
      schema:
        type: string
        pattern: ^00-[0-9a-f]{32}-[0-9a-f]{16}-[0-9a-f]{2}$
      description: |
        W3C Trace Context header used for distributed tracing. Format: 00-<trace-id>-<span-id>-<trace-flags>. Example: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-0
...
    NovationGUID:
      name: novationGUID
      in: path
      description: The GUID of the Novation.
      required: true
      schema:
        type: string
        format: uuid
...
    QPageNumber:
      name: page
      in: query
      required: false
      schema:
        type: integer
      description: Page index for given request (start from 0, default is 0)
    QPageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
      description: Page size for given request (default is 100)
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: X-Authorization
      in: header
    AzureAD:
      type: oauth2
      description: |
        OAuth2 client credentials via Microsoft Entra ID. Callers acquire a token using `scope={backend-api-application-id}/.default`. The API then enforces the `PDA_ACCESS` application role from the JWT `roles` claim.
      flows:
        clientCredentials:
          tokenUrl: https://login.microsoftonline.com/{tenant-id}/oauth2/v2.0/token
          scopes:
            '{backend-api-application-id}/.default': |
              Request app permissions granted to the client app for this API (including `PDA_ACCESS`).